Extracts summary statistics from a raster file based on the boundaries of a shapefile

The function extract_enm2

extract_enm2(enm, geoshape, id0, id, th = NA, multipar = FALSE, ncores = 1)

Arguments

enm
input raster data (usually a species distribution model, SDM)

geoshape
input shapefile with boundaries of geometries to summarize the raster data for

id0
part of 'geoshape@data' that holds the names of the geopolitical units to focus on (polygons of interest)

id
subset in geoshape identifier table to iterate through: a vector of names, may need to match this to a particular part of geoshape with id0

th
threshold of suitability used to calculate RSA. If not provided, the RSA is not calculated.

multipar
use multicore processors to perform the function in parallel (default is false)

ncores
tell the function to use a set number of cores if multipar = TRUE and defaults to ncores=1

Value

Dataframe (tibble) that contains columns that include summary statistics from the input.